requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
httpx>=0.25.0
ollama>=0.1.7
pypdf2>=3.0.1
python-docx>=0.8.11
